modify the type; pcm_length and such are no longer i32

+++ sound_programming/src/lib.rs
@@ -112,23 +112,23 @@ pub fn safe_Hanning_window(w_slice: &mut [c_double])
   	}
 }
 
-pub fn wave_read_16bit_mono_safer2(path : &str) -> (&[f64], i32, i32, i32){
+pub fn wave_read_16bit_mono_safer2(path : &str) -> (&[f64], i32, i32, usize){
 	unsafe{
 		let mut pcm : MONO_PCM = mem::uninitialized();
 		wave_read_16bit_mono(&mut pcm, to_c_str(path));
 		let pcm_slice = from_raw_parts(pcm.s, pcm.length as usize);
-		return (pcm_slice, pcm.fs, pcm.bits, pcm.length);
+		return (pcm_slice, pcm.fs, pcm.bits, pcm.length as usize);
 	}
 }
 
 #[allow(non_snake_case)]
-pub fn wave_read_16bit_stereo_safer2(path : &str) -> (&[f64], &[f64], i32, i32, i32){
+pub fn wave_read_16bit_stereo_safer2(path : &str) -> (&[f64], &[f64], i32, i32, usize){
 	unsafe{
 		let mut pcm : STEREO_PCM = mem::uninitialized();
 		wave_read_16bit_stereo(&mut pcm, to_c_str(path));
 		let pcm_sliceL = from_raw_parts(pcm.sL, pcm.length as usize);
 		let pcm_sliceR = from_raw_parts(pcm.sR, pcm.length as usize);
-		return (pcm_sliceL, pcm_sliceR, pcm.fs, pcm.bits, pcm.length);
+		return (pcm_sliceL, pcm_sliceR, pcm.fs, pcm.bits, pcm.length as usize);
 	}
 }
 
@@ -138,11 +138,11 @@ pub fn to_c_str(a: &str) -> *mut i8 {
 }
 
 #[allow(non_snake_case)]
-pub fn wave_write_16bit_mono_safer2(path: &str, x : (&mut [f64], i32, i32, i32) ){
+pub fn wave_write_16bit_mono_safer2(path: &str, x : (&mut [f64], i32, i32, usize) ){
 	let mut pcm1 : MONO_PCM = MONO_PCM{
 			fs : x.1,
 			bits : x.2,
-			length : x.3,
+			length : x.3 as i32,
 			s : x.0.as_mut_ptr()
 		};
 	unsafe{
@@ -151,11 +151,11 @@ pub fn wave_write_16bit_mono_safer2(path: &str, x : (&mut [f64], i32, i32, i32) 
 }
 
 #[allow(non_snake_case)]
-pub fn wave_write_16bit_stereo_safer2(path: &str, x : (&mut [f64], &mut [f64], i32, i32, i32) ){
+pub fn wave_write_16bit_stereo_safer2(path: &str, x : (&mut [f64], &mut [f64], i32, i32, usize) ){
 	let mut pcm1 : STEREO_PCM = STEREO_PCM{
 			fs : x.2,
 			bits : x.3,
-			length : x.4,
+			length : x.4 as i32,
 			sL : x.0.as_mut_ptr(),
 			sR : x.1.as_mut_ptr()
 		};
